Removing src structure.
