+++ pybl/raincell/__init__.py
@@ -1,2 +1,1 @@
-from .cell import *
 from .rcimodel import *
