+++ pybl/raincell/rcimodel.py
@@ -30,12 +30,12 @@ class IConstantRCI(Protocol):
 
 class GammaRCIModel(IConstantRCI):
     @staticmethod
-    @nb.njit  # type: ignore
+    @nb.njit
     def get_f1(sigmax_mux: float) -> float:
         return (sigmax_mux + 1.0) / sigmax_mux
 
     @staticmethod
-    @nb.njit  # type: ignore
+    @nb.njit
     def get_f2(sigmax_mux: float) -> float:
         x2 = sigmax_mux**2
         return (x2 + 3.0 * sigmax_mux + 2.0) / x2
@@ -50,12 +50,12 @@ class GammaRCIModel(IConstantRCI):
 
 class ExponentialRCIModel(IConstantRCI):
     @staticmethod
-    @nb.njit  # type: ignore
+    @nb.njit
     def get_f1(sigmax_mux: float) -> float:
         return 2.0
 
     @staticmethod
-    @nb.njit  # type: ignore
+    @nb.njit
     def get_f2(sigmax_mux: float) -> float:
         return 6.0
 
@@ -71,14 +71,14 @@ class ExponentialRCIModel(IConstantRCI):
 
 class WeibullRCIModel(IConstantRCI):
     @staticmethod
-    @nb.njit  # type: ignore
+    @nb.njit
     def get_f1(sigmax_mux: float) -> float:
         ex2: float = math.gamma(1.0 + 2.0 / sigmax_mux)
         ex: float = math.gamma(1.0 + 1.0 / sigmax_mux)
         return ex2 / ex**2
 
     @staticmethod
-    @nb.njit  # type: ignore
+    @nb.njit
     def get_f2(sigmax_mux: float) -> float:
         ex3: float = math.gamma(1.0 + 3.0 / sigmax_mux)
         ex: float = math.gamma(1.0 + 1.0 / sigmax_mux)
@@ -98,12 +98,12 @@ class WeibullRCIModel(IConstantRCI):
 
 class ParetoRCIModel(IConstantRCI):
     @staticmethod
-    @nb.njit  # type: ignore
+    @nb.njit
     def get_f1(sigmax_mux: float) -> float:
         return (sigmax_mux - 1.0) ** 2 / sigmax_mux / (sigmax_mux - 2.0)
 
     @staticmethod
-    @nb.njit  # type: ignore
+    @nb.njit
     def get_f2(sigmax_mux: float) -> float:
         return (sigmax_mux - 1.0) ** 3 / sigmax_mux**2 / (sigmax_mux - 3.0)
 
@@ -121,12 +121,12 @@ class ParetoRCIModel(IConstantRCI):
 
 class GPRCIModel(IConstantRCI):
     @staticmethod
-    @nb.njit  # type: ignore
+    @nb.njit
     def get_f1(sigmax_mux: float) -> float:
         return 2.0 * (1.0 - sigmax_mux) / (1.0 * 2.0 * sigmax_mux)
 
     @staticmethod
-    @nb.njit  # type: ignore
+    @nb.njit
     def get_f2(sigmax_mux: float) -> float:
         return (
             6.0
