--- src/pybl/fitting/__init__.py
@@ -1,1 +1,0 @@
-from .fitter import *
