--- src/pybl/timeseries/__init__.py
@@ -1,1 +1,0 @@
-from .intensity_mrle import *
